Added new features and methods

+++ src/FileHandler.java
@@ -2,10 +2,10 @@ import java.io.*;
 import java.util.*;
 import org.json.simple.*;
 import org.json.simple.parser.*;
-
 public class FileHandler {
     private static final String FILE_NAME = "recipes.json";
 
+    // Метод для сохранения списка рецептов в JSON-файл
     public static void saveToFile(List<Recipe> recipes) {
         JSONArray array = new JSONArray();
 
@@ -22,10 +22,10 @@ public class FileHandler {
             obj.put("favorite", r.isFavorite());
             obj.put("creationDate", r.getCreatedDate());
             obj.put("inPlan", r.isInPlan());
-
             array.add(obj);
         }
 
+        // Запись JSON-массива в файл
         try (FileWriter file = new FileWriter(FILE_NAME)) {
             file.write(array.toJSONString());
             file.flush();
@@ -34,17 +34,20 @@ public class FileHandler {
         }
     }
 
+    // Метод для загрузки рецептов из JSON-файла
     public static List<Recipe> loadFromFile() {
-        List<Recipe> recipes = new ArrayList<>();
-
+        List<Recipe> recipes = new ArrayList<>();  // Список рецептов, который вернём
         try (FileReader reader = new FileReader(FILE_NAME)) {
+            // Считываем содержимое файла
             JSONArray array = (JSONArray) new JSONParser().parse(reader);
 
+            // Проходим по каждому объекту массива
             for (Object obj : array) {
                 JSONObject o = (JSONObject) obj;
 
+                // Создаём объект рецепта на основе данных JSON
                 Recipe r = new Recipe(
-                        (int) o.get("id"),
+                        ((Long) o.get("id")).intValue(),
                         (String) o.get("name"),
                         (String) o.get("description"),
                         (List<String>) o.get("ingredients"),
@@ -56,7 +59,6 @@ public class FileHandler {
                         (String) o.get("creationDate"),
                         (Boolean) o.get("inPlan")
                 );
-
                 recipes.add(r);
             }
         } catch (Exception e) {
